Implement transformation that discover logsumexp pattern (#913)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/913

With PPL models it is quite common to expect the summation pattern `log( exp(a) + exp(b) + exp(c) ...)` , for example in logistic regression with more than 2 categories. This pattern can instead be replaced with a more stable `logsumexp(a,b,c, ...)` operation. In BMG this transformation leads to reduction in the number of nodes and edges, which could translate to inference speedup.

Design Choice:
There are two ways we could handle this transformation, we can make the transformations not depend on each other or we can make it ordered. Former would lead to duplicated transformation efforts. For example, in this case we would need to detect that there is a chain of summations which multiary addition already does for us. Therefore for this version, I decided to go with the latter, where we assume an order to the transformations i.e. to carry out `LogSumExpFixer` we need to do `MultiaryOperatorFixer` first.  However, we should make a todo somewhere that if a user selectively disables transformations, we need to alert them that it could effectively disable other transformations that depend on it.

Fixing Criteria:
As a simple case, a log expression is fixable if operand is a `MultiAdditionNode`. Additionally, each input of this `MultiAdditionNode` is an `ExpNode`.

Test:
The new test case shows a simple case of how the rewrite works. We start with the following model:

{F631027695}

After rewrite we get the following:

{F631027529}

Since we ensure `MultiaryOperatorFixer` is carried out before `LogSumExpFixer`, the following graph does not meet the fixing criteria, therefore it is not transformed to use `logsumexp`.

{F631028191}

Reviewed By: ericlippert

Differential Revision: D29640106

fbshipit-source-id: 39fb82e64d36433570147435e5d54a959a2b3b40