Performance test for LogSumExp transformation
Summary:
This test checks for performance improvement when transforming graphs to use `Logsumexp` primitive.
When using the following metrics to measure performance from the BMG performance report,
we observe a reduction in the size of the graph:
a. number of nodes - for this test 205 → 104 (49%)
b. number of edges - for this test 303 → 202 (33%)

However, wrt to runtime, we find that the transformed model performs worse than the original model (e.g. 2147ms to 3923 ms). `Logsumexp` is only an accuracy hack, the increase in the number of arithmetic operations when using the hack overshadows the reduction in number of nodes/edges leading to poor runtime performance.
See https://fb.quip.com/xHiDA7jWnJQk#JUDABAybmTt for more details.

Reviewed By: horizon-blue

Differential Revision: D29701337

fbshipit-source-id: a0a07f69346f01e44fa3a7fd6517b1202b3abf38