Split categorical-with-logits apart from categorical
Summary: BMG supports some distributions taking a stochastic log-probability as an input, but not others (yet; we should consider adding support).  To represent this in the graph accumulator I have followed the pattern I've used for other distributions; the categorical distribution node now has two versions, one for regular probabilities and one for log probabilities, and an abstract base class for the code they have in common.

Differential Revision: D29179387

fbshipit-source-id: 5ef68b02af431c064bba775be5b2e95169297b18