Split categorical-with-logits apart from categorical (#920)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/920

BMG supports some distributions taking a stochastic log-probability as an input, but not others (yet; we should consider adding support).  To represent this in the graph accumulator I have followed the pattern I've used for other distributions; the categorical distribution node now has two versions, one for regular probabilities and one for log probabilities, and an abstract base class for the code they have in common.

Reviewed By: feynmanliang

Differential Revision: D29179387

fbshipit-source-id: f0e69cd8c70d51c703b9568da6eb475349967634