Add support for generating constant simplex nodes to compiler
Summary:
BMG makes a distinction between a matrix of probabilities and a simplex in its type system. (That is, a matrix of probabilities where each column adds to 1.0).  When we support categorical distributions we will need the ability to mark a constant node as a simplex that is an input to a categorical.

This diff adds:

* A node class for constant simplexes
* Factory methods to add the node to the graph builder
* Code to transform a constant simplex node in to C++ code, Python code, a BMG node, and so on.

I have not yet implemented the code which converts an untyped constant to a constant simplex; that will come in a later diff, when we can exercise that code in a test.

Reviewed By: feynmanliang

Differential Revision: D29508747

fbshipit-source-id: 5e95bd71bb0e5593fdd276a7a87f67eb2fe14a99