Refactor NMC out of Graph
Summary: We will refactor NMC to delegate taking steps to random-variable-type-specific "Steppers", keeping NMC more generic. These steppers will need to access NMC-provided methods, so it is best if NMC is not an encapsulated class of Graph anymore. Besides, this agrees with the direction of freeing up Graph from algorithm dependencies (the rationale being that Graph is merely a data structure that should not know about, or depend on, algorithms manipulating it).

Reviewed By: yucenli

Differential Revision: D29624033

fbshipit-source-id: aaf70fa813d8566113c1ac297ececff1893b8c7b