Introduce `NMC::revertibly_set_and_propagate` and `NMC::revert_set_and_propagate`
Summary:
Introduce NMC operations for setting a node to a new value and update its immediate deterministic descendants in a revertible manner, as well as for reverting it.

At this point, NMC is operating as a sort of "BMG graph enriched for NMC operations". So it makes sense for it to have this type of state manipulation operation.

This is used in the NMC scalar stepper to make it a bit more readable, but will probably be reusable in other functions as refactoring progresses.

Reviewed By: yucenli

Differential Revision: D29893639

fbshipit-source-id: 53d136c55d65a4c495fb68f4a61b00c26e02da52