Let single-site steppers indicate which type of node they apply to
Summary: NMC hard-coded the selection of single-site stepper based on the type of node. This made it harder to re-use the algorithm for different types of steppers (it would require changing its main loop code). So I extended the interface for single-site steppers with a method that tests a node for applicability of the stepper. Now the NMC main loop only needs to go over available steppers and check which one applies to the current target node. This way, we can simple add new single-site steppers to the algorithm and it will work without modification.

Reviewed By: yucenli

Differential Revision: D29653950

fbshipit-source-id: 1e75b37454b279210021bb5ac26f1bcf70d0f8fc