Replace BMG binary multiplication node with multiary multiplication node (#931)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/931

This is a code improvement change similar to the previous diff in the stack, no functionality changes.

Currently Beanstalk considers binary and multiary multiplication as two different constructs. Multiary multiplication is used only when transforming chains of binary multiplication. Since binary multiplication is a case of multiary multiplication, we can unify these and have a single addition node that supports n operands.

Graph Changes:
I have removed MultiMultiplicationNode and modified MultiplicationNode to take a list of inputs. This node now does same computation (i.e. size, support, support_size etc) as a binary node when the `len(inputs)` is 2. So `_edge_lable` for this node becomes `_numbered_or_left_right` which returns `_left_right` when its a binary addition.

Fix changes:
Fixes that check for multiplication node, additionally assert that the length is 2 and use index 0 and 1 instead of left and right for input nodes.

Reviewed By: feynmanliang

Differential Revision: D29943314

fbshipit-source-id: 370eda5a5a34ebb591c62b913277ad06b22a10c2