Replace BMG binary addition node with multiary addition node (#921)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/921

This is a code improvement change, no functionality changes.

Currently Beanstalk considers binary and multiary addition as two different constructs. Multiary addition is used only when transforming chains of binary addition. Since binary addition is a case of multiary addition, we can unify these and have a single addition node that supports `n` operands.

Graph Changes:
I have removed `MultiAdditionNode` and modified `AdditionNode` to take a list of inputs. This node now does same computation (i.e. `size`, `support`, `support_size` etc) as a binary node when the `len(inputs)` is `2`. So `_edge_lable` for this node becomes `_numbered` which returns `_left_right` when its a binary addition.

Fix changes:
Fixes that check for addition node, additionally assert that the length is 2 and use index 0 and 1 instead of left and right for input nodes.
`LogSumExpFixer` now checks for `AdditionNode` instead of `MultiAdditionNode`; so now we can handle `log(exp(a) + exp(b))` case as well.

Reviewed By: feynmanliang, ericlippert, horizon-blue

Differential Revision: D29765412

fbshipit-source-id: 125ffbca568387fa30c58e14506d81ac233c31b1