Create a "natural choice" operator in BMG (#934)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/934

BMG has an `IF_THEN_ELSE` operator which has a Boolean first parent, and then two additional parents of the same type.  I am here adding a `CHOICE` operator which has a natural first parent and n additional parents of the same type. We presume that the support of the first parent is 0 to no more than n-1; if that turns out to be false then inference can crash. (Just as it can if we try to do a stochastic index into a non-existing element of a matrix.)

If we ever support bounded naturals as a type in BMG, we can improve the type safety of this operation; until then we will just insert a check in `eval`.

Since indexing into a matrix is essentially this same operation, why not just use that?  Because a matrix can only be two-dimensional in BMG. That means that to have a stochastic choice amongst, say, ten 2x2 arrays, we would need to index into a 10x2x2 array and we cannot represent such an array.

My intention is to use this operator for representing stochastic choices between random variables when compiling models to BMG; those changes will come in a later diff.

While adding this operator I fixed some typos and made error handling in `IfThenElse` more consistent with the other operators. I also simplified some needlessly repetitious `IfThenElse` code.

Reviewed By: feynmanliang, yucenli

Differential Revision: D29833681

fbshipit-source-id: 75f4cfa6754e549fa630a715df1de57103512cae