Stochastic control flows now generate choice operator in compiler (#937)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/937

As noted in previous diffs, in BMG we were representing a stochastic choice amongst several values as an index into an array of atomic values. This has several problems; first off, it seems like a hack rather than a purpose-built solution. And second, an array must be two-dimensional in BMG, which makes it hard to use this trick to choose from amongst matrix values.  (This is not an academic concern; our version of the CLARA model which uses categoricals chooses from amongst several possible simplex-valued samples.)

In the previous diff in this stack I added a `CHOICE` operator to BMG; in this diff I:

* add a Python version of that operator in the compiler
* ensure that its type analysis matches the requirements of the BMG node
* update the logic that handles "switch on natural" to generate `CHOICE(c, ...)` rather than `INDEX(TO_MATRIX(...), c)`
* this causes the output of the GMM test to change
* this now allows the categorical CLARA model to compile without error

Reviewed By: feynmanliang

Differential Revision: D29861609

fbshipit-source-id: f91214b83c4b0fb641eb9e88dae7bad9eff0915d