Rename inf_types to node_types
Summary:
When I originally designed the lattice typing system to analyze graphs I referred to the node's type as its "inf type", since we were finding the infimum type in the lattice that the node was compatible with.  Several refactorings later, I no longer need to call out that detail so I've been removing uses of the term from the codebase.

In this diff I just rename the parameter on `to_dot` that controls whether node types are displayed.

Reviewed By: feynmanliang

Differential Revision: D29949299

fbshipit-source-id: 9ab41cf0ba5bbcc66aec8900895f09377dd7b31a