Implement transformation that discover beta-bernoulli conjugate pattern
Summary:
This transformation rewrites graphs with Bernoulli likelihood and Beta prior into a graph with an updated Beta prior. Currently this transformation is disabled by default.

Since this is a conjugate pair, we analytically update the prior parameters `Beta(alpha, beta)` using observations to get the posterior parameters `Beta(alpha', beta')`. Once we update the parameters, we delete the observed samples from the graph. This greatly decreases the number of nodes, the number of edges in the graph, and the Bayesian update is reduced to parameter update which can lead to performance wins during inference.

Graph Rewrite (Fixing) Criteria:
We will carry out transformation if we see the following:

1. We encounter a bernoulli node
2. This bernoulli node has an incoming edge which is a sample coming from a Beta
3. The Beta contains two incoming edges which are both constant nodes
4. The beta sample should be queried

Design Choice:
For this transformation, we identify patterns starting at `Bernoulli` node and check for the fixing conditions. An alternate method would be to start with observations and identify a beta-bernoulli pattern. The former would check for the pattern once whereas the latter would check for the same pattern for every observation.

Test:
`beta_binomial_basic_test` shows a simple case of how this rewrite works. We start with the following model

{F637054942}

And after rewrite we get the following

{F637054975}

Test cases that will be tested in future diffs:
1. Models with additional operations on the Beta/Bernoulli RVs
2. Models where Beta parameters themselves are RVs. Eg. Beta(Normal(0.1, 1.0), 5.0) . Note that this would also require change in the graph rewriting logic.
3. Models that use mixture of conjugate priors - e.g. `theta` = `0.25 Beta(1, 1) + 0.75 Beta(5,5)`

Differential Revision: D30001791

fbshipit-source-id: 2d71cf2a9c3ebaee405da32f6dfdd18995199216