Correctness testcases for beta-bernoulli conjugate pair transformation
Summary:
Beta-Bernoulli-based conjugacy transformations are applied when the hyper-parameters are constants.

The series of diffs following this captures different model scenarios as testcases. In each of the test case, we check that the original and transformed model produce samples from same distribution using KS test. Where applicable, we also check that the transformed graph matches the expected transformation.

In this diff we check the first simple case when the model has a bernoulli likelihood with 4 observations and a beta prior whose parameters are constants. In this case transformation should be carried out and we call this the purely conjugate case.

Differential Revision: D30088388

fbshipit-source-id: b82b3cde0a921e29cb7222a2c25132a9549bc18f