Transformation test when there are operations on Bernoulli node in a Beta-Bernoulli model (#942)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/942

In this diff we check the case when the model has a bernoulli likelihood with 4 observations and a beta prior with constant parameters. Further, in the model we have some operation on samples from Bernoulli node.

In this case transformation should be carried out and it needs to retain the bernoulli node and sample nodes that are used in the downstream operations. We check if the original and the transformed model produce similar samples and that the graphs after transformation match.

Graph before transformation:

{F638945161}

Graph after transformation:

{F638945229}

Reviewed By: feynmanliang

Differential Revision: D30115391

fbshipit-source-id: 025ab884c477c8a187ac41f69ec58cfcbc595843