Transformation test when the hyperparameter of a Beta-Bernoulli model is a RV (#943)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/943

In this diff we check the case when the model has a bernoulli likelihood with 4 observations and a beta prior whose parameters take values from a random variable.

In this case transformation should `not` be carried out and it needs to produce the same graph as the original model.
We check if the original and the transformed model produce similar samples and that the graphs remain the same even when the `BetaBernoulliConjguateFixer` is enabled.

Graph illustrating this case:

{F638943842}

Reviewed By: feynmanliang

Differential Revision: D30115407

fbshipit-source-id: 865dffcf575df6ac584d1793a2d557ca8437a39a