Transformation test when a constant given by calling a non-random_variable function is added to the hyperparameter of a Beta-Bernoulli model (#941)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/941

In this diff we check the case when the model has a bernoulli likelihood with 4 observations and a beta prior. Beta's `alpha` parameter has a constant given by calling a non-random_variable function added to the hyperparameter. In this case transformation should be carried out and we check if the original and the transformed model produce similar samples and that the graphs after transformation match.

Reviewed By: feynmanliang

Differential Revision: D30115313

fbshipit-source-id: 9e16b261982f5ea51c51d0fe9d306020efcdaca0