Remove duplicate functionality from dirichlet beta proposer (#945)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/945

create_proposer_dirichlet_beta was being used for bot obtaining a proposer and computing the log prob, even though NMC's compute_log_prob already does that. This has the benefit of removing duplicate code and also giving a more clearly defined function to create_proposer_dirichlet_beta. This is part of making NMC Dirichlet Beta stepper more and more similar to the scalar one so they can be as unified as possible.

Reviewed By: yucenli

Differential Revision: D29996395

fbshipit-source-id: 42b1c6d8f25a8d20b9a00d3c3635fb8b78a4a686