Remove duplicate functionality from dirichlet beta proposer
Summary: create_proposer_dirichlet_beta was being used for bot obtaining a proposer and computing the log prob, even though NMC's compute_log_prob already does that. This has the benefit of removing duplicate code and also giving a more clearly defined function to create_proposer_dirichlet_beta. This is part of making NMC Dirichlet Beta stepper more and more similar to the scalar one so they can be as unified as possible.

Differential Revision: D29996395

fbshipit-source-id: 7f855c23f5bcdd655b5eb2444f115c2caa897cb1