Remove grad1 = 1 hack from NMC Dirichlet steppers (#946)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/946

Previously the `StochasticOperator::gradient_log_prob` method depended on a node's `grad1` field being equal to 1 in order to decide how to compute the gradient. However this wouldn't work for matrix-valued nodes which use `Grad1` instead, so client code needed to unnecessarily set `grad1` to 1 just to get that to work. However `gradient_log_prob` was fixed in D29341532 (https://github.com/facebookresearch/beanmachine/commit/2260a113a62c93e5b2a19ab7fe702914dbe41e27) and this is no longer necessary, and in this diff we are removing the now unnecessary setting of `grad1` to 1 in the NMC Dirichlet steppers.

Reviewed By: yucenli

Differential Revision: D29996823

fbshipit-source-id: b47c3b3d56a6acb810f66adb3dd186f67c63d545