Remove duplicate functionality from dirichlet beta proposer (#945)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/945

create_proposer_dirichlet_beta was being used for bot obtaining a proposer and computing the log prob, even though NMC's compute_log_prob already does that. This has the benefit of removing duplicate code and also giving a more clearly defined function to create_proposer_dirichlet_beta. This is part of making NMC Dirichlet Beta stepper more and more similar to the scalar one so they can be as unified as possible.

Reviewed By: yucenli

Differential Revision: D29996395

fbshipit-source-id: e13262af072aa7a8a1527e9c8bc321a0e9954cf8