Refactor BetaBernoulliConjguateFixer to separate prior and likelihood checks
Summary:
Beta distribution is a conjugate prior to a number of likelihood distributions. Beta is a conjugate prior to Bernoulli, Binomial, Negative Binomial and Geometric distributions. The graph pattern check for the prior remains the same for all of the pairs. The only difference across these pairs is in the way we update the beta parameters.

In this diff I separate the conjugate transformation into a `BetaPriorFixer` and `BetaBernoulliConjguateFixer` class such that we can reuse most of the code across different distributions and avoid most code duplication.

In upcoming diffs, I plan to add conjugate transformation support for beta-binomial pair and its test cases. Since we dont support Geometric/Negative Binomial distributions in BMG yet, I will leave this as future work.

Reviewed By: horizon-blue

Differential Revision: D30193254

fbshipit-source-id: d333d8a124caef2045f26a1db834135f749908eb