Implement transformation that discover and transforms beta-binomial conjugate pattern
Summary:
This transformation rewrites graphs with Binomial likelihood and Beta prior into a graph with an updated Beta prior. Currently this transformation is disabled by default.

Since this is a conjugate pair, we analytically update the prior parameters `Beta(alpha, beta)` using observations to get the posterior parameters `Beta(alpha', beta')`. Once we update the parameters, we delete the observed samples from the graph. This greatly decreases the number of nodes, the number of edges in the graph, and the Bayesian update is reduced to parameter update which can lead to performance wins during inference.

Graph Rewrite Criteria:
We will carry out transformation if we see the following:
1. We encounter a Binomial node
2. This binomial node has an incoming edge which is a sample coming from a Beta
3. The Beta contains two incoming edges which are both constant nodes
4. The beta sample should be queried

Design Choice:
We have to repeat the graph transformations for each conjugate pair separately, since we do not have equivalences in the type checker yet to hoist bernoulli/binomial etc to a general type and carry out one transformation instead. Also, since we may want to selectively toggle on/off certain conjugate pair transformations, we introduce a new class for each of them.

Test:
`fix_beta_binomial_basic_test` shows a simple case of how this rewrite works. We start with the following model

{F640412771}

After rewrite we get the following:

{F640412842}

In the upcoming diffs, I plan to refactor the conjugate test case models into `beanmachine/ppl/examples/conjugate_models` and use one test case per transformation to test for all of these models.

Reviewed By: horizon-blue

Differential Revision: D30193342

fbshipit-source-id: 9816c2360c397b8404679ef9fa3b5933b3989a85