Implement transformation that discover and transforms normal-normal conjugate pattern (#963)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/963

This transformation rewrites graphs with Normal likelihood with fixed `sigma` and a Normal prior for `mu` into a graph with an updated Normal. Currently this transformation is disabled by default.

Since this is a conjugate pair, we analytically update the prior parameters `Normal(mu, sigma)` using observations to get the posterior parameters `Normal(mu', sigma')`.

Graph Rewrite Criteria:
We will carry out transformation if we see the following:
1. We encounter a Normal node

2. This normal nodes mu parameter has an incoming edge which is a sample coming from a normal
3. This normal nodes sigma parameter is a constant node

4. The Normal node for mu contains two incoming edges which are both constant nodes

5. The Normal sample for mu should be queried

Reviewed By: horizon-blue

Differential Revision: D30314170

fbshipit-source-id: 9ceecc04710b5827305cc44d61ad3e455ee58065