Normal-Normal template to test original and conjugate prior-based transformed model (#962)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/962

Given a BaselineModel of a normal likelihood with fixed variance and normal prior for `mu`, we want Beanstalk compiler to identify that it is an instance of conjugate prior and produce a transformed model that looks like TransformedModel.

In this test case, I have explicitly added the transformation computation (`transformed_mu_p`, `transformed_sigma_p`) as a part of the model. Currently this test checks (using KS) if these two models produce samples from the same distribution. Once we have implemented the transformation, we could additionally check that the graphs produced from them are syntactically same.

Reviewed By: feynmanliang

Differential Revision: D29961022

fbshipit-source-id: be2ff0dc8e66987b6f7bdd61cd32ef5679c0eac5