Unify clear_gradients
Summary: Scalar and Dirichlet-Beta steppers both used clearing of gradients but the code was different because one was for scalar nodes and the other for matrix nodes. I've created some utilities in the NMC class to provide a uniform interface for clearing gradients, using it instead, making the code a bit closer to identical on the path to unify them.

Differential Revision: D30327946

fbshipit-source-id: 9b75ead17d5a0a23fee9168557272050b14353ad