Change dirichlet beta step code to use the "true value" Y of the node
Summary:
In the dirichlet beta stepper, the true value of the node is Y (a 2-element matrix representing a Dirichlet sample), but the MH code is expressed in terms of X, its first element (the second element is 1 - X). This prevents unification with a standard stepper because a standard stepper manipulates the true value of the node.

So this diff starts changing the code to use the true value Y instead of X. After this diff, the code uses Y, with the except of the proposal distribution, which still works in terms of X (it samples a new value for X and evaluates the log prob of X). The next step is to write a proposal distribution that samples Y and evaluates the log prob of Y, which will allow us to write the MH code purely in terms of Y.

Reviewed By: yucenli

Differential Revision: D30325391

fbshipit-source-id: 312d9d9e8458985d9e38c5c64f1e4b810c6a1a48