Remove need to pass det_affected_nodes and sto_affected_nodes along with target node.
Summary:
The NMC class computes the sets of deterministic and stochastic affected nodes for each unobserved support node in the BMG. These sets are organized in their respective vectors, and we know to which node they correspond by knowing that the i-th unobserved support node has the i-th set of deterministic affected nodes and the i-th set of stochastic affected nodes.

However, if we represent a node solely by `Node *`, we have no way to obtain its unobserved support node index without an actual linear search over the array of unobserved support nodes. In order to avoid that, lots of methods taking a target node represented by `Node*` had to also receive its sets deterministic and stochastic affected nodes, making the code clunky and error-prone.

We fix it here by introducing a new vector in the NMC class that maps the node index (`node->index`), which is its index in the variables in the BMG, to its unobserved support node index. We then modify several methods to look that up and find the sets of deterministic and stochastic affected nodes by themselves instead of requiring them as parameters. This results in code that reads a lot more easily.

Reviewed By: yucenli

Differential Revision: D30391194

fbshipit-source-id: 02c6c35938ca0af17cff26e6598335d0bb6763ec