Expose target accept prob configuration API
Summary: Per OpenTeams' request, it would be quite convenient if we let users configure the target accept probability for NUTS and HMC :D. All this diff does is to connect the wires and make the API publicly available. A small unit test is also added to ensure that the target probs work as expected.

Differential Revision: D30735168

fbshipit-source-id: a6ee2519164a7c65ff901af0c6e4dd5d7ed9e739