Expose target accept prob configuration API (#980)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/980

Per OpenTeams' request, it would be quite convenient if we let users configure the target accept probability for NUTS and HMC :D. All this diff does is to connect the wires and make the API publicly available. A small unit test is also added to ensure that the target probs work as expected.

Reviewed By: feynmanliang

Differential Revision: D30735168

fbshipit-source-id: 289a8ed198ee05d8d3a09f0924557f1954af8d84