Compute tensor sizes associated with graph nodes (#989)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/989

We need to know precisely what tensor size is associated with each node in an accumulated graph. Why? Because we need this information:

(1) to implement multi-dimensional samples and other arithmetic when compiling to BMG, and

(2) to give sensible error messages when we cannot, and

(3) to compute the support of graph nodes when one is used as an argument to an RV

This is essentially the same task as assigning a lattice type to every graph node, and we have the same problem: a graph might have a path more than 1000 edges long, so a recursive solution will blow up Python's small recursion limit.

I therefore am going to remove the recursive size computation from graph nodes themselves -- and besides, that's not a concern of the node -- and move that computation to a class that can perform this computation nonrecursively, caching the results as it goes.

In this diff I create that class; in upcoming diffs I will similarly move support computation to its own class, since it has the same problem.

Reviewed By: wtaha

Differential Revision: D29948346

fbshipit-source-id: 071aaefc1adfa8df48707d6c6951cfac7a915fd1