Move SetOfTensors class to own module
Summary:
The SetOfTensors class is wrong, wrong, wrong; turns out, it only works for one-dimensional tensors.

I'm going to fix it, but first I'm going to refactor it into its own module, as I've been intending to do for some time.

Since this functionality might be of use outside the compiler proper I'm putting it in the utils directory.

Reviewed By: wtaha

Differential Revision: D30082749

fbshipit-source-id: 6721e6ad264e40f630bef31655d7116e5708b9f2