Abstract MH from NMC
Summary:
In this diff we finally abstract the MH functionality from the previous NMC implementation. Essentially, the previous NMC becomes an MH class that is independent of any NMC-specific code (namely, code about which types are supported by NMC, and which steppers to use for NMC).

Now the NMC class is a small class sub-classing MH and specifying the NMC-specific aspects mentioned above. This new NMC class has the same interface as the previous one, so code using it remains the same.

Differential Revision: D30769916

fbshipit-source-id: dec206e5b6d3ee06448c1d21650bce10459eef2b