Fix bugs in SetOfTensors (#991)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/991

My SetOfTensors class now does what it says on the tin: makes a deduplicated set of tensors.

I have not implemented all the features you'd typically find in a set; I only need construction, length, iteration and membership so that's all I've done.

Because tensors do not implement hashing or equality in the manner necessary to do set membership, I've created a wrapper around a set of tuples or values where the tuples/values have the same "shape" as the corresponding tensor.  I do this because tuples do implement hashing and equality, so I can make a set of tensor equivalents and convert them back to tensors on demand.

Reviewed By: wtaha

Differential Revision: D30115798

fbshipit-source-id: c5969682fdfc16232447441eed051e5b7b9b7ecf