Start implementing nonrecursive graph node support computation
Summary:
To handle stochastic control flows of the form

    some_rv(some_stochastic_expression_with_finite_support)

we need to enumerate that finite support during graph accumulation. However, there could be thousands of nodes that are ancestors of that stochastic expression, and the support might depend on all of them. We cannot compute the support recursively because if there is a long path through the graph, we blow Python's recursion limit.

We already have a base class which implements non-recursive computation and memoization of a per-node property; this is the base class used by the lattice typer and node sizer. We therefore extend it once more to assign supports to nodes.

We represent a support by a set of tensors. Much of the time this will just be an ordinary deduplicated set of tensors; there are three cases where it is not:

* If the support of a node is infinite -- say, a sample from a normal -- then we represent it as a special Infinite value.
* Some stochastic nodes will have finite but large support. We can estimate the size of the support before we compute it, and if it is likely to be too large, we skip the computation and represent it as a special "too big" value
* Nodes for which we have not yet implemented computation of the support will be represented by an "unknown" value.

In this diff I just implement computing support for:

* samples from infinite-support distributions
* samples from Bernoulli draws of arbitrary tensor size
* construction of stochastic tensors

We'll add the rest in subsequent diffs.

We have an existing implementation of support computation via instance methods which is (1) recursive and (2) frequently wrong. Once all necessary nodes have their support computation added to the new class, I will delete the instance methods.

Reviewed By: wtaha

Differential Revision: D30050828

fbshipit-source-id: 4912d75ff5df7166b6b759cabb2fd36c872c3b1d