Abstract MH from NMC
Summary:
In this diff we finally abstract the MH functionality from the previous NMC implementation. Essentially, the previous NMC becomes an MH class that is independent of any NMC-specific code (namely, code about which types are supported by NMC, and which steppers to use for NMC).

Now the NMC class is a small class sub-classing MH and specifying the NMC-specific aspects mentioned above. This new NMC class has the same interface as the previous one, so code using it remains the same.

Reviewed By: yucenli

Differential Revision: D30769916

fbshipit-source-id: e6571d25f8adae89343a4d75c5a762714e8ec484