Implement prototype of model devectorizer
Summary:
By design, we lack facilities in BMG to run inference on vectorized models. That is, BMG for example supports sampling from Bernoulli(0.5) to get a Boolean, and it supports independent-identical-distribution sampling from Bernoulli(0.5) to get a vector of independent Booleans, but it does NOT support sampling from Bernoulli([0.25, 0.75]) to get a vector of independent but *non-identical* samples.

However we would like to be able to compile such models to BMG if possible. Doing so will allow us to test whether we still get a performance improvement from BMG for such models.

Doing so will require a series of changes. First, we need to know when a model even is vectorized, and when the vectors are of the right shape to permit compilation to BMG. (Recall BMG only supports atomic values and 2-d matrices of atomic values.) My recent improvements to the tensor size computation on graphs will help solve that problem.

Second, we need to find an appropriate transformation from the vectorized graph to a less vectorized graph that BMG can handle. We will need to have custom code for each distribution type, so there will be many changes required to get there.

This diff simply implements a proof of concept that we can in fact do so successfully in principle. It implements a transformation ONLY on Bernoulli distributions where the input is a `Size([2])` tensor.

As you can see from the before and after graphs, this works!  Note though that there is a small inefficiency: a rewrite which introduces a constant index into a constant tensor does not simply add the constant result to the graph. I will add that optimization in a later diff.

This transformation correctly handles queries, which can be on any node, but does not correctly handle observations. If we have an observation of a `Size([2])` sample, we will end up with a graph where we have an observation of a ToMatrix, which is wrong. BMG only permits observations of sample nodes.  In an upcoming diff I will add a feature to the rewriter to handle this case.

BEFORE:

{F639574372}

AFTER:

{F639574527}

Reviewed By: wtaha

Differential Revision: D30165910

fbshipit-source-id: d51331c522cf75f4d592bf743b637065c45d5e42