Add constant folding optimization for index operations (#997)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/997

When a rewriter generates an indexing operation it is possible to end up with constants for both the collection and the index; in that case we can write a constant folding optimization.  In this diff I refactored the index rewriter to make the logic a bit easier to follow by extracting the code for vector index to its own method, and then added a case for constants for both operands.

Graph before optimization implemented:

{F639574527}

after optimization:

{F639592295}

Reviewed By: wtaha

Differential Revision: D30167970

fbshipit-source-id: d7063927a1a636572d828165f4c8eb62614666b3