Model unvectorizer prototype now supports observations
Summary:
My model unvectorizer prototype still only supports one kind of model: a Bernoulli with two inputs instead of one.  In this diff I add the ability to have a two-value observation on such a Bernoulli; the unvectorizer rewrites it into two observations of two samples and deletes the original sample and observation.

I had assertions that observations were always observing samples, which is an invariant of BMG. However, due to the incremental nature of the rewriting process we end up with a temporary situation where an observation observes a stochastic tensor. I've removed those assertions.

Reviewed By: wtaha

Differential Revision: D30500846

fbshipit-source-id: b722615f04d19f9f7169794eb73bc8b28071610f