Extend devectorizer prototype to support arbitrary 1-d Bernoulli tensors (#1000)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1000

I am continuing to experiment with my prototype devectorizer; previously we only supported the test case of a Bernoulli with a probability of size 2. We now implement Bernoulli with probability of any size >= 2 so long as it is single-dimensional. Supporting two-dimensional probabilities, or distributions other than Bernoulli, is not yet implemented.

Reviewed By: wtaha

Differential Revision: D30522236

fbshipit-source-id: 2a454ad40d1be42fa0b567b620c86e2f2025551b