implement support cache for discrete RVs
Summary:
Records the max cardinality of a random_variable based on calling `update_support`. Only valid for finite support dists (Bernoulli and Categoricals right now).  Right now this will only be used in an enumerate method (upcoming diff), and so is not invoked each time `update_graph` is called.

Needed for discrete enumeration of RVs

Reviewed By: rodrigodesalvobraz

Differential Revision: D30390183

fbshipit-source-id: e05651da2708988d91ac3b99a65b0b09608d22d9