Start eliminating _required_columns property (#1004)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1004

I wish to remove the `size` property from all graph nodes, which means first eliminating all code which consumes this property.  The `_required_columns` property on a Dirichlet node uses this property; this does not need to be a property of the node.  We can compute it when we need it in the lattice typer or sizer. (And also the property is misnamed, as it has the semantics of required *rows*.  Might as well delete it instead of fixing the name.)

This diff begins the process of removing usages of this property, starting with the code generators.

Reviewed By: wtaha

Differential Revision: D30586326

fbshipit-source-id: 02e5bdd31fad378340555f2a3f5876bd97bbdbd4