Finish removing _required_columns property from Dirichlet node
Summary:
I have finished removing the incorrectly named, badly located and poorly implemented `_required_columns` property from `DirichletNode`.  The name was wrong because it actually computed required rows; it was badly located because the input requirement is a concern of the BMG requirements module, not the node, and it was poorly implemented because it tried to handle input types unsupported by BMG.

All these problems are now fixed; the logic is no longer in the node so the badly named method no longer exists. The code is now relocated to the lattice type and requirements modules. The implementation now makes no attempt to handle unsupported scenarios; in the case where a model has an unsupported tensor shape as the Dirichlet input we give up on attempting to determine its type and default to "1x1 simplex", and then give an error during requirements checking.

This diff furthers the larger objective of removing the `size` property from node classes.

Reviewed By: wtaha

Differential Revision: D30592028

fbshipit-source-id: 2b242326e839cc7021924f0cc7a708484470796a