Continue implementing support computation (#1008)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1008

In order to build a graph for control flows of the form `some_rv(some_other_rv())` we compute the support -- the set of all possible values -- for  `some_other_rv` and then determine the value of `some_rv` for each possibility.  This means that we must be able to compute the support of any node representing a value in the graph.

My intention in this refactoring is to move that code out of the node classes themselves and into a class similar to the one which computes node types. We require that this algorithm be non-recursive.

In this diff I implement support computation for the majority of the unary and binary operator nodes. This is quite straightforward because for all of those operators we can compute the support of an operator `foo(x)` or `foo(x, y)` by computing all possible combinations of the operands, and then computing the value of the operator for those combinations.

I have constructed a mapping from node type to a function which computes the value of the operator; we compute the Cartesian product of the operand supports and pass them to the appropriate function to compute the possible values.

In upcoming diffs I will remove the remaining callers of the existing instance methods and delete the instance methods.

Reviewed By: wtaha

Differential Revision: D30608810

fbshipit-source-id: 76a29cb35e65b5b822c61f2ffbe8f978cdef7f50